+++ src/superannotate/__init__.py
@@ -2,7 +2,7 @@ import logging
 import os
 import sys
 
-__version__ = "4.6.1"
+__version__ = "4.6.2dev1"
 
 
 os.environ.update({"sa_version": __version__})